+++ app/javascript/oxalates/FoodListError.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import './food-list-error.scss';
+import './FoodListError.scss';
 
 export const FoodListError = ({ error }) => (
   <li className="food-list-item error">{error}</li>
