+++ app/javascript/oxalates/FoodListError.scss
@@ -1,0 +1,5 @@
+@import "./FoodListItem";
+
+.error {
+  color: #f60;
+}
