+++ app/javascript/oxalates/FoodListItem.scss
