+++ app/javascript/oxalates/FoodListItem.test.jsx
@@ -3,9 +3,9 @@ import Enzyme, { shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import renderer from 'react-test-renderer';
 
-import foods from '../../../__mocks__/foods.json';
-import { FoodListItem } from './food-list-item';
-import lists from '../../../__mocks__/lists.json';
+import foods from './__mocks__/foods.json';
+import { FoodListItem } from './FoodListItem';
+import lists from './__mocks__/lists.json';
 
 Enzyme.configure({ adapter: new Adapter() });
