+++ app/javascript/oxalates/FoodListLoading.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import './food-list-loading.scss';
+import './FoodListLoading.scss';
 
 export const FoodListLoading = () => (
   <li className="food-list-item">Loading foods...</li>
