+++ app/javascript/oxalates/FoodListLoading.scss
@@ -1,0 +1,1 @@
+@import "./FoodListItem";
