+++ app/javascript/oxalates/ListHeader.test.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import lists from '../../../__mocks__/lists.json';
-import { ListHeader } from './list-header';
+import lists from './__mocks__/lists.json';
+import { ListHeader } from './ListHeader';
 
 test('ListHeader renders the All list title', () => {
   const component = renderer.create(
