+++ app/javascript/oxalates/ListMenuItem.scss
