+++ app/javascript/oxalates/ListMenuItem.test.jsx
@@ -3,8 +3,8 @@ import Enzyme, { shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import renderer from 'react-test-renderer';
 
-import lists from '../../../__mocks__/lists.json';
-import { ListMenuItem } from './list-menu-item';
+import lists from './__mocks__/lists.json';
+import { ListMenuItem } from './ListMenuItem';
 
 Enzyme.configure({ adapter: new Adapter() });
