+++ app/javascript/oxalates/actions/selectActiveList.js
@@ -1,4 +1,4 @@
-import { fetchFoods } from './fetch-foods';
+import { fetchFoods } from './fetchFoods';
 
 export const SELECT_ACTIVE_LIST_ID = 'SELECT_ACTIVE_LIST_ID';
 export const selectActiveListId = activeListId => {
