--- app/javascript/oxalates/food-list-error.scss
@@ -1,5 +1,0 @@
-@import "./food-list-item";
-
-.error {
-  color: #f60;
-}
