--- app/javascript/oxalates/food-list-loading.scss
@@ -1,1 +1,0 @@
-@import "./food-list-item";
