+++ app/javascript/oxalates/foodPropType.js
